requirements.txt
absl-py==0.9.0
astor==0.8.1
beautifulsoup4==4.9.1
boto3==1.14.7
botocore==1.17.7
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
docutils==0.15.2
feedparser==5.2.1
gast==0.3.3
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.1
jmespath==0.10.0
joblib==0.15.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lxml==4.5.1
Markdown==3.2.2
nltk==3.5
numpy==1.19.0
protobuf==3.12.2
python-dateutil==2.8.1
python-docx==0.8.10
regex==2020.6.8
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
tensorboard==1.12.2
tensorflow-gpu==1.12.0
tensorflow-hub==0.7.0
termcolor==1.1.0
torch==1.2.0
tqdm==4.46.1
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
torch>=1.8.0
transformers>=4.5.0
pandas>=1.2.0
scikit-learn>=0.24.0
python==3.6
